% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{box_plots}
\alias{box_plots}
\title{Boxplots}
\usage{
box_plots(df)
}
\arguments{
\item{df}{a data.frame returned from MTT::import}
}
\value{
The boxplots of the raw data
}
\description{
Boxplots
}
\examples{
df <- import("./data/testfile1.xml")
box_plots(df)
}
