% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{conc_default}
\alias{conc_default}
\title{Returns default design for concentrations}
\usage{
conc_default()
}
\value{
The default data.frame used for concentration
}
\description{
Returns default design for concentrations
}
