% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import data}
\usage{
import(path, names = names_default(), conc = conc_default())
}
\arguments{
\item{path}{defines the path to the file which should be imported}

\item{names}{is a data.frame which defines the names of the wells. This argument is optional. The default value can be checked using names_default()}

\item{conc}{is a data.frame which defines the concentrations within the wells. This argument is optional. The default value can be checked using conc_default()}
}
\value{
data.frame containg the imported data with the corresponding names and conc.
}
\description{
Import data
}
\examples{
df <- import("./data/testfile1.xml")
df
}
