% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld50.R
\name{ld50_fit}
\alias{ld50_fit}
\title{Import data}
\usage{
ld50_fit(df)
}
\arguments{
\item{df}{a data.frame returned from MTT::import}
}
\value{
The LD50 and the quality of the fit for each substance
}
\description{
Import data
}
\examples{
df <- import("./data/testfile1.xml")
ld50(df)
}
