% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ld50plot}
\alias{ld50plot}
\title{LD50 plots}
\usage{
ld50plot(df)
}
\arguments{
\item{df}{a data.frame returned from MTT::import}
}
\value{
The LD50 plots
}
\description{
LD50 plots
}
\examples{
df <- import("./data/testfile1.xml")
ld50plot(df)
}
