% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{names_default}
\alias{names_default}
\title{Returns default design for substances}
\usage{
names_default()
}
\value{
The default data.frame used for naming
}
\description{
Returns default design for substances
}
