% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{removeOutlier}
\alias{removeOutlier}
\title{remove outliers}
\usage{
removeOutlier(df, l, dep)
}
\arguments{
\item{df}{a data.frame from which the outliers should be removed}

\item{l}{a list containing column indices which defines the different groups}

\item{dep}{the column index which defines the dependent variable}
}
\description{
remove outliers
}
