% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf.R
\name{report_pdf}
\alias{report_pdf}
\title{Creates report file}
\usage{
report_pdf(boxplots, ld50_values, aov_res, file_name)
}
\arguments{
\item{boxplots}{result of box_plots()}

\item{ld50_values}{result of ld50()}

\item{aov_res}{result of stats()}

\item{filename}{name of excel file}
}
\value{
Nothing is returned. Just called for its sideeffects
}
\description{
Creates report file
}
