% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel.R
\name{save}
\alias{save}
\title{Saves data}
\usage{
save(ic_50List, file_name)
}
\arguments{
\item{ic_50List}{the list result of a call to ic50.}

\item{file_name}{the name of the excel file where the results should be saved.}
}
\description{
Saves data
}
\examples{
path <- system.file("data", package = "MTT")
df <- read.csv(paste0(path, "/ExampleData.txt"))
res <- ic50(df, "abs", "conc", "names", "neg", "pos")
save(res, "results.xlsx")
}
