% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Statistics of raw data}
\usage{
stats(df)
}
\arguments{
\item{df}{a data.frame returned from MTT::import}
}
\value{
The aov results of the raw data
}
\description{
Statistics of raw data
}
\examples{
df <- import("./data/testfile1.xml")
stats(df)
}
