% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{XML2DF}
\alias{XML2DF}
\title{Conversion of xml to data.frame}
\usage{
XML2DF(string, names, conc)
}
\arguments{
\item{string}{is a xml file as string}

\item{names}{is a data.frame which defines the names of the wells. This argument is optional. The default value can be checked using names_default()}

\item{conc}{is a data.frame which defines the concentrations within the wells. This argument is optional. The default value can be checked using conc_default()}
}
\value{
data.frame containg the imported data with the corresponding names and conc.
}
\description{
Conversion of xml to data.frame
}
