% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf.R
\name{report_pdf}
\alias{report_pdf}
\title{Creates report file}
\usage{
report_pdf(ic_50List, text, file_name)
}
\arguments{
\item{ic_50List}{the list result of a call to ic50.}

\item{text}{the variable text which has to be inserted in the document}

\item{file_name}{the name of the excel file where the results should be saved.}
}
\description{
Creates report file
}
\examples{
path <- system.file("data", package = "MTT")
df <- read.csv(paste0(path, "/ExampleData.txt"))
res <- ic50(df, "abs", "conc", "names", "neg", "pos")
report_pdf(res, "blabla", "report.pdf")
}
