% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf.R
\name{report_pdf}
\alias{report_pdf}
\title{Creates report file}
\usage{
report_pdf(boxplots, ld50_values, ld50fit, stat_aov, text, file_name)
}
\arguments{
\item{boxplots}{result of box_plots()}

\item{ld50_values}{result of ld50()}

\item{ld50fit}{result of ld50 fit summary}

\item{text}{the variable text which has to be inserted in the document}

\item{filename}{name of excel file}
}
\value{
Nothing is returned. Just called for its sideeffects
}
\description{
Creates report file
}
