% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel.R
\name{save}
\alias{save}
\title{Saves data}
\usage{
save(ld50plots, boxplots, ld50_values, aov_res, file_name)
}
\arguments{
\item{ld50plots}{result of ld50plot()}

\item{boxplots}{result of box_plots()}

\item{ld50_values}{result of ld50()}

\item{aov_res}{result of stats()}

\item{filename}{name of excel file}
}
\value{
The aov results of the raw data
}
\description{
Saves data
}
\examples{
df <- import("./data/testfile1.xml")
stats(df)
}
