% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{XML2DF}
\alias{XML2DF}
\title{Conversion of xml to data.frame}
\usage{
XML2DF(string)
}
\arguments{
\item{string}{is a xml file as string}
}
\value{
data.frame containg the imported data with the corresponding names and conc.
}
\description{
Conversion of xml to data.frame
}
